requirements.txt
langdetect
transformers
torch
nltk
beautifulsoup4
datasets
panda
biopython
sacremoses
sacrebleu
